requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-client-cache==0.7.3
aiosignal==1.2.0
aiosqlite==0.17.0
async-lru==1.0.3
async-property==0.2.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
discord.py==2.0.1
emoji==2.0.0
frozenlist==1.3.1
idna==3.3
itsdangerous==2.1.2
marshmallow==3.17.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
multidict==6.0.2
oauthlib==3.2.0
orjson==3.7.12
packaging==21.3
Pillow==9.2.0
pycares==4.2.2
pycparser==2.21
pyparsing==3.0.9
pyshorteners==1.0.1
python-dotenv==0.20.0
python-forge==18.6.0
pytimeparse==1.1.8
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.10.0
typing_extensions==4.3.0
url-normalize==1.4.3
urllib3==1.26.11
yarl==1.8.1
